Initial work on classes. Put on hold for now
