+++ lib/Monocle_Asner.php
@@ -1,0 +1,13 @@
+<?php
+namespace Scanner;
+
+/**
+ * Created by PhpStorm.
+ * User: JamesHarland
+ * Date: 09/08/2017
+ * Time: 17:26
+ */
+class Asner extends Monocle
+{
+
+}
